# classic-version/vue/src/api/models.ts
// This file is auto-generated by @hey-api/openapi-ts

export type Body_api_login_api_login_post = {
  grant_type?: string | null
  username: string
  password: string
  scope?: string
  client_id?: string | null
  client_secret?: string | null
}

export type Body_set_user_avatar_api_v1_users__user_id__avatar_post = {
  avatar: Blob | File
}

export type HTTPValidationError = {
  detail?: Array<ValidationError>
}

export type NewPassword = {
  token: string
  password: string
}

export type Token = {
  access_token: string
  token_type?: string
}

export type UserCreate = {
  email: string
  username: string
  first_name?: string | null
  last_name?: string | null
  password: string
}

export type UserAndToken = {
  user: UserOut
  token: Token
}

export type UserOut = {
  email: string
  username: string
  first_name?: string | null
  last_name?: string | null
  id: number
  is_active: boolean
  avatar_url?: string | null
}

export type UserUpdate = {
  email?: string | null
  password?: string | null
  username?: string | null
  first_name?: string | null
  last_name?: string | null
}

export type ValidationError = {
  loc: Array<string | number>
  msg: string
  type: string
}

export type $OpenApiTs = {
  '/api/login': {
    post: {
      req: {
        formData: Body_api_login_api_login_post
      }
      res: {
        /**
         * Successful Response
         */
        200: UserAndToken
      }
    }
  }
  '/api/recover-password/{email}': {
    post: {
      req: {
        email: string
      }
      res: {
        /**
         * Successful Response
         */
        204: void
      }
    }
  }
  '/api/reset-password/': {
    post: {
      req: {
        requestBody: NewPassword
      }
      res: {
        /**
         * Successful Response
         */
        201: UserOut
      }
    }
  }
  '/api/v1/users/': {
    post: {
      req: {
        requestBody: UserCreate
      }
      res: {
        /**
         * Successful Response
         */
        201: UserAndToken
      }
    }
    get: {
      req: {
        offset?: number
        query?: string
      }
      res: {
        /**
         * Successful Response
         */
        200: Array<UserOut>
      }
    }
  }
  '/api/v1/users/me': {
    get: {
      res: {
        /**
         * Successful Response
         */
        200: UserOut
      }
    }
  }
  '/api/v1/users/{user_id}': {
    get: {
      req: {
        userId: number
      }
      res: {
        /**
         * Successful Response
         */
        200: UserOut
      }
    }
    patch: {
      req: {
        requestBody: UserUpdate
        userId: number
      }
      res: {
        /**
         * Successful Response
         */
        201: UserOut
      }
    }
    delete: {
      req: {
        userId: number
      }
      res: {
        /**
         * Successful Response
         */
        204: void
      }
    }
  }
  '/api/v1/users/{user_id}/avatar': {
    post: {
      req: {
        formData: Body_set_user_avatar_api_v1_users__user_id__avatar_post
        userId: number
      }
      res: {
        /**
         * Successful Response
         */
        201: UserOut
      }
    }
  }
}